Moved utilities methods to bespoke places
- checks to validation
- comparisons into own file
- dataset calculations to dataset_properties
